support: move http functions into client class
Being able to hold on to a curl handle is useful for maintaining
connection state like cookies. I'd like to use cookies to store some
connection state with the name server.

The most natural way (that I can think of) to encapsulate this is with a client
object which libtsuba can hold onto.

Signed-off-by: Tyler Hunt <thunt@katanagraph.com>